bpf: drop tunnel traffic not marked as decrypted wireguard traffic
This commits adds a feature that lets users drop any pod-to-pod traffic
that hasn't been encrypted. It can be enabled by setting
--enable-encryption-strict-mode-ingress, or the respective helm value.
For now, this feature will only be available when Wireguard and
tunneling are enabled as well.

This new strict ingress encryption mode is unrelated to the already
existing strict mode encryption feature. To avoid any confusion, the
next commit will deprecate the exisiting feature flags and helm values
for the strict mode encryption and rename them to have a distinct
separation between ingress and egress strict mode encryption.

Signed-off-by: Robin Gögge <r.goegge@isovalent.com>