agent:overlay: fallback to available underlay protocol automatically
Prior to this commit, it was possible to select a disabled underlay
protocol, causing potential MTU calculation errors. For example, in
an IPv6-only cluster, the defaul value for underlayProtocol is IPv4,
and nothing prevents the agent from running with this incorrect information.
That leads to error in computing the MTU, accounting for 50B of overhead
rather than the 70B for the IPv6 underlay case.

The above problems lead to fragmentation, and in some configuration
to packet drops and no pod-to-pod connectivity. Let's automatically
adjust the parameter as needed. It is not up to us here to check whether
both underlays are disabled: that is done in the `initEnv` daemon function.

Another option of this PR would have been to simply error out in the
`newConfig` method, but that would have required ALL IPv6-only users
to suddendly change their cluster configs on upgrades, due to the default
`IPv4` protocol being disabled. For this reason, we choose to change its
value at runtime.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>