cilium, bigtcp: BIG TCP for tunnels
Don't fail hard when BIG TCP is requested in tunneling mode. Rather
probe tso_max_size on the tunnel interface to check whether the kernel
support for BIG TCP for UDP tunnels has landed. If so, set gso_max_size
and co on the cilium_vxlan or cilium_geneve netdev as well.

Signed-off-by: Maxim Mikityanskiy <maxim@isovalent.com>