ci:wireguard: enable Host Firewall in native routing e2e tests
This enabled Host Firewall in the `wireguard-3` config, so that we make
sure that host-related packets go through the host firewall hook in
`cilium_host` when WireGuard is enabled.

This will be tested only in scheduled runs, given host-firewall related
CLI tests are skipped in `/ci-l7`, which is basically all WireGuard configs.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>