cli: clustermesh: use ca bundle to connect clusters
Make the CLI use ca bundle instead of the deprecated per cluster
key/certs. It introduce a new step to configure CA bundles on all
clusters before the actual connection to avoid any errors in our CI.

We also skip the following warning: `CEP was deleted externally or there
is a significant delay on the CEP informer` triggered on the agent when
a clustermesh-apiserver restart in kvstoremesh mode. This is getting
triggered when we start configuring the CA bundle that essentially
restart every Cilium components.

Signed-off-by: Arthur Outhenin-Chalandre <git@mrfreezeex.fr>